+++ cmd/bundle/run.go
@@ -8,7 +8,6 @@ import (
 	"os"
 
 	"github.com/databricks/cli/bundle"
-	"github.com/databricks/cli/bundle/config/engine"
 	"github.com/databricks/cli/bundle/env"
 	"github.com/databricks/cli/bundle/phases"
 	"github.com/databricks/cli/bundle/resources"
@@ -133,11 +132,6 @@ Example usage:
 	cmd.Flags().BoolVar(&restart, "restart", false, "Restart the run if it is already running.")
 
 	cmd.RunE = func(cmd *cobra.Command, args []string) error {
-		engine, err := engine.FromEnv(cmd.Context())
-		if err != nil {
-			return err
-		}
-
 		b, err := utils.ProcessBundle(cmd, utils.ProcessOptions{
 			SkipInitialize: true,
 		})
@@ -172,7 +166,11 @@ Example usage:
 			return executeScript(content, cmd, b)
 		}
 
-		ctx, stateDesc := statemgmt.PullResourcesState(ctx, b, statemgmt.AlwaysPull(true), engine)
+		requiredEngine, err := utils.ResolveEngineSetting(ctx, b)
+		if err != nil {
+			return err
+		}
+		ctx, stateDesc := statemgmt.PullResourcesState(ctx, b, statemgmt.AlwaysPull(true), requiredEngine)
 		if logdiag.HasError(ctx) {
 			return root.ErrAlreadyPrinted
 		}